add Dockerfile
add a Dockerfile to descripe how the app can be packaged into a docker
container. This requires that the app is build using Gradle first so
there is a Java JAR file that can be deployed inside the Docker
container.
Also add a .dockerignore file to ignore all files not needed for the
build.