only display character when it could be loaded
up until now we showed character information even when no character with
the given id could be loaded. That resulted in some layout elements
getting layed out with no content. We now do not draw layout elements
for a character that could not be found. As a side effect this will also
ensure that no empty elements are drawn when the server is slow to send
the character data.