add basic dialog showing the skill check result
For rolling dice a new service was introduced. It is likely that we will
implement a number of dice rolls in the application so implementing all
in one centrel service makes sense if we ever have to touch how dice
rolls are done in the future. At the moment dice rolls are implemented
using Math.random().